+++ packages/cloud-cognitive/scripts/generate/templates/DISPLAY_NAME.stories.js
@@ -9,19 +9,18 @@ import React from 'react';
 // TODO: import action to handle events if required.
 // import { action } from '@storybook/addon-actions';
 
-import { pkg } from '../../settings';
-import { getStorybookPrefix } from '../../../config';
-import { prepareStory } from '../../global/js/utils/story-helper';
+import {
+  getStoryTitle,
+  prepareStory,
+} from '../../global/js/utils/story-helper';
 
 import { DISPLAY_NAME } from '.';
 import mdx from './DISPLAY_NAME.mdx';
 
 import styles from './_storybook-styles.scss';
 
-const storybookPrefix = getStorybookPrefix(pkg, DISPLAY_NAME.displayName);
-
 export default {
-  title: `${storybookPrefix}/${DISPLAY_NAME.displayName}`,
+  title: getStoryTitle(DISPLAY_NAME.displayName),
   component: DISPLAY_NAME,
   // TODO: Define argTypes for props not represented by standard JS types.
   // argTypes: {