+++ packages/cloud-cognitive/src/components/CreateTearsheet/preview-components/MultiStepTearsheet.js
@@ -1,3 +1,4 @@
+/* eslint-disable react/prop-types */
 /**
  * Copyright IBM Corp. 2021, 2021
  *
@@ -22,7 +23,16 @@ import { CreateTearsheetStep } from '../CreateTearsheetStep';
 
 const blockClass = `${pkg.prefix}--tearsheet-create-multi-step`;
 
-export const MultiStepTearsheet = () => {
+export const MultiStepTearsheet = ({
+  backButtonText,
+  cancelButtonText,
+  className,
+  description,
+  label,
+  nextButtonText,
+  submitButtonText,
+  title,
+}) => {
   const [simulatedDelay] = useState(750);
   const [open, setOpen] = useState(false);
   const [shouldReject, setShouldReject] = useState(false);
@@ -53,14 +63,14 @@ export const MultiStepTearsheet = () => {
         {open ? 'Close CreateTearsheet' : 'Open CreateTearsheet'}
       </Button>
       <CreateTearsheet
-        className={blockClass}
-        submitButtonText="Create"
-        cancelButtonText="Cancel"
-        backButtonText="Back"
-        nextButtonText="Next"
-        description="Specify details for the new topic you want to create"
-        label="This is the label of the multi step tearsheet"
-        title="Create topic"
+        label={label}
+        className={cx(blockClass, className)}
+        submitButtonText={submitButtonText}
+        cancelButtonText={cancelButtonText}
+        backButtonText={backButtonText}
+        nextButtonText={nextButtonText}
+        description={description}
+        title={title}
         open={open}
         onClose={clearCreateData}
         onRequestSubmit={() =>
@@ -100,7 +110,7 @@ export const MultiStepTearsheet = () => {
             connection information, so make it something easy to recognize.
           </p>
           <TextInput
-            labelText="Topic name*"
+            labelText="Topic name"
             id="tearsheet-multi-step-story-text-input-multi-step-1"
             value={stepOneTextInputValue}
             placeholder="Enter topic name"
@@ -119,14 +129,14 @@ export const MultiStepTearsheet = () => {
             }}
           />
           <TextInput
-            labelText="Topic description"
+            labelText="Topic description (optional)"
             id="tearsheet-multi-step-story-text-input-multi-step-1-input-2"
             value={topicDescriptionValue}
             placeholder="Enter topic description"
             onChange={(event) => setTopicDescriptionValue(event.target.value)}
           />
           <TextInput
-            labelText="Topic version"
+            labelText="Topic version (optional)"
             id="tearsheet-multi-step-story-text-input-multi-step-1-input-3"
             value={topicVersionValue}
             placeholder="Enter topic version"